refactor schema en multiples archivos
