+++ server/schema/Mutation.js
@@ -1,0 +1,5 @@
+module.exports = `
+type Mutation {
+  talkAdd(input: TalkAddInput!) : TalkAddPayload!
+}
+`;
