+++ server/schema/Query.js
@@ -1,0 +1,6 @@
+module.exports = `
+type Query {
+  viewer: User!
+  node(id: ID!): Node
+}
+`;
