+++ server/schema/constants.js
@@ -1,0 +1,3 @@
+module.exports = {
+  CONNECTION_ARGS: 'after: String, first: Int, before: String, last: Int'
+};
